+++ paper/paper.bib
@@ -6,6 +6,7 @@
   pages={110030},
   year={2020},
   publisher={Elsevier},
+  doi={10.1016/j.rser.2020.110030},
   url={https://doi.org/10.1016/j.rser.2020.110030}
 }
 @article{ching2009national,
@@ -17,6 +18,7 @@
   pages={1157--1168},
   year={2009},
   publisher={American Meteorological Society},
+  doi={10.1175/2009BAMS2675.1},
   url={https://doi.org/10.1175/2009BAMS2675.1}
 }
 @inproceedings{mills2015introduction,
@@ -34,6 +36,7 @@
   number={9},
   pages={1907--1924},
   year={2018},
+  doi={10.1175/BAMS-D-16-0236.1},
   url={https://doi.org/10.1175/BAMS-D-16-0236.1}
 }
 
@@ -45,6 +48,7 @@
   number={65},
   pages={3541},
   year={2021},
+  doi={10.21105/joss.03541},
   url={https://doi.org/10.21105/joss.03541}
 }
 
