### 2024-04-05 e01086e
update quickstarter with dagworks platform instructions
