### 2024-04-09 6c1ffe2
update quickstarter with graphviz instructions, clearer wording
