+++ .github/workflows/build.yml
@@ -41,7 +41,9 @@ jobs:
             pip install pytest-cov
             pytest --cov=naturf/ --cov-report=xml
         - name: Upload coverage to Codecov
-          uses: codecov/codecov-action@v3
+          uses: codecov/codecov-action@v4
+          env:
+            CODECOV_TOKEN: ${{ secrets.CODECOV_TOKEN }}
           with:
             file: ./coverage.xml
             fail_ci_if_error: true
