### 2024-05-02 #6712 @editorialbot
[REVIEW]: naturf: a package for generating urban parameters for numerical weather modeling
**Submitting author:** @levisweetbreu (<a href="http://orcid.org/0000-0003-3325-8086">Levi Sweet-Breu</a>)
**Repository:** https://github.com/IMMM-SFA/naturf.git
**Branch with paper.md** (empty if default branch):
**Version:** v1.0.3
