### 2024-05-02 IMMM-SFA/naturf#164 @praneethd7
[JOSS] Glossary of terms in the user guide
Hello! Thank you for making this package open-source. I was going through the [user guide](https://immm-sfa.github.io/naturf/user_guide.html) and I noticed there were a lot of terms used. It would be nice to have a glossary of terms. Also, I noticed some inconsistencies such as the building height represented by both $zh$ and $z_h$.

Tagging the review thread here:
 #openjournals/joss-reviews#6712
