### 2024-05-02 #6712 @praneethd7
@levisweetbreu The manuscript has the following statement:

> NATURF was used to demonstrate that simulated new developments in the Chicago Loop neighborhood in Chicago Illinois, USA affect temperature and energy use both in the new developments and the preexisting neighborhoods (Allen-Dumas et al., 2020).

However, I did not find any mention of the package or the word `NATURF` in the manuscript [@allen2020impacts](https://www.sciencedirect.com/science/article/pii/S136403212030321X?via%3Dihub). The way it is currently written sounds like the package existed before the publication.
