### 2024-05-02 #6712 @cheginit
@praneethd7 Thanks for starting the review!

I just have one note. It'd be better to create issues in `naturf`'s repo for each comment that you have, so the authors can address them, and you can close them whenever they address it. This makes it much easier to keep track of issues. Whenever you open an issue, please make sure to reference `openjournals/joss-reviews#6712` so it shows up here, and I can follow the progress.

Thank you!
