### 2024-05-03 IMMM-SFA/naturf#165 @levisweetbreu
An early version of NATURF was used to calculate the urban parameters used in [@allen2020impacts](https://www.sciencedirect.com/science/article/pii/S136403212030321X?via%3Dihub), so I have modified the manuscript to say:

> A preliminary version of NATURF was used to calculate urban parameters and demonstrate that simulated new developments in the Chicago Loop neighborhood in Chicago Illinois, USA affect temperature and energy use both in the new developments and the preexisting neighborhoods (Allen-Dumas et al., 2020).

Thank you for your comment!
