### 2024-05-03 12107e9
synthesize terms in equations #164
