+++ docs/source/user_guide.rst
@@ -58,7 +58,7 @@ Glossary of Terms
 * :math:`\overline{\lambda_s}`: Average height-to-width ratio for given building and neighbors (unitless)
 * :math:`\overline{z_h}`: Average building height for given building and neighbors (`m`)
 * `SVF`: Sky-view factor (unitless)
-* `W`: Average distance between given building and neighbors (`m`)
+* :math:`\overline{W}`: Average distance between given building and neighbors (`m`)
 * :math:`rl_{go}`: Grimmond & Oke rougness length (`m`)
 * :math:`dh_{go}`: Grimmond & Oke displacement height (`m`)
 * :math:`rl_r`: Raupach roughness length (`m`)
@@ -85,7 +85,7 @@ Frontal Area Density (1-60)
 Frontal area density is the frontal area at a certain height increment divided by the total plan area. **naturf** calculates frontal area density from the four cardinal directions (east, north, west, south) and at 5 meter increments from ground level to 75 meters. Parameters 1-15 represent the north, parameters 16-30 represent the west, parameters 31-45 represent the south, and parameters 46-60 represent the east. For instance, parameter 1 gives the north-facing wall area for each building and its neighbors divided by the total plan area. [Burian2003]_ Eq. 14
 
 .. math::
-  FAD = \frac{FA}{TPA}
+  FAD = \frac{A_{f}}{A_{P}}
 
 where `FAD` is Frontal area density; `FA` is the frontal area of the wall from the current direction and height level in :math:`m^2`; `TPA` is the total plan area in :math:`m^2`.
 
@@ -95,7 +95,7 @@ Plan Area Density (61-75)
 Plan area density is the ratio of building plan area to the total plan area, calculated in 5 meter increments from ground level to 75 meters. **naturf** projects the building footprint vertically to the building height, meaning plan area density is the same at every vertical level. [Burian2003]_ Eq. 7
 
 .. math::
-  PAD = \frac{BPA}{TPA}
+  \lambda_p = \frac{A_{b}}{A_{P}}
 
 where, `PAD` is the plan area density; `BPA` is the building plan area in :math:`m^2`; `TPA` is the total plan area in :math:`m^2`.
 
@@ -126,7 +126,7 @@ The average height of all buildings within the total plan area weighted by the t
 
 .. math::
 
-  AWMH = \frac{\Sigma{A_i zh_i}}{\Sigma{A_i}}
+  AWMH = \frac{\Sigma{A_{b} z_{h}}}{\Sigma{A_{b}}
 
 where, `AWMH` is the area weighted mean height in m; `A_i` is the current building plan area in `m^2`; `zh_i` is the current building height in m.
 
@@ -141,7 +141,7 @@ Frontal Area Index (96-99)
 Frontal area index is the ratio of the entire frontal area of a building to the total plan area. **naturf** calculates the frontal area index from the four cardinal directions. Because buildings often do not face a cardinal direction head on, **naturf** uses the average alongwind and crosswind distance from the current building centroid to all other building centroids for the total plan area. [Burian2003]_ Eq. 12
 
 .. math::
[…]
