### 2024-05-03 cf28449
remove equation captions #164
