+++ docs/source/user_guide.rst
@@ -87,7 +87,6 @@ Frontal area density is the frontal area at a certain height increment divided b
 .. math::
   FAD = \frac{A_{f}}{A_{P}}
 
-where `FAD` is Frontal area density; `FA` is the frontal area of the wall from the current direction and height level in :math:`m^2`; `TPA` is the total plan area in :math:`m^2`.
 
 Plan Area Density (61-75)
 ^^^^^^^^^^^^^^^^^^^^^^^^^
@@ -97,7 +96,6 @@ Plan area density is the ratio of building plan area to the total plan area, cal
 .. math::
   \lambda_p = \frac{A_{b}}{A_{P}}
 
-where, `PAD` is the plan area density; `BPA` is the building plan area in :math:`m^2`; `TPA` is the total plan area in :math:`m^2`.
 
 Rooftop Area Density (76-90)
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^
@@ -128,7 +126,6 @@ The average height of all buildings within the total plan area weighted by the t
 
   AWMH = \frac{\Sigma{A_{b} z_{h}}}{\Sigma{A_{b}}
 
-where, `AWMH` is the area weighted mean height in m; `A_i` is the current building plan area in `m^2`; `zh_i` is the current building height in m.
 
 Building Surface Area to Plan Area Ratio (95)
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
@@ -143,7 +140,6 @@ Frontal area index is the ratio of the entire frontal area of a building to the 
 .. math::
   \lambda_{f} = \frac{A_{F}`}{A_{P}}
 
-where, `FAI` is frontal area index; `FA` is the frontal area of the wall from the current direction in :math:`m^2`; `TPA` is the total plan area in :math:`m^2`.
 
 Complete Aspect Ratio (100)
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^
@@ -153,7 +149,6 @@ The ratio of building surface area and exposed ground area to the total plan are
 .. math::
   CAR = \frac{BSA + (A_{P} - A_{b)}{A_{P}}
 
-where, `BSA` is the building surface area in :math:`m^2`; `BPA` is the building plan area in :math:`m^2`; `TPA` is the total plan area in :math:`m^2`.
 
 Height-to-Width Ratio (101)
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^
@@ -163,7 +158,6 @@ The ratio of the building height to the street width. **naturf** generalizes thi
 .. math::
   \overline{\lambda_s} = \frac{\overline{z_h}}{\overline{W}}
 
-where, :math:`\overline{\lambda_s}` is the average height-to-width ratio; :math:`\overline{z_h}` is the average building height in m; :math:`\overline{W}` is the average distance between buildings.
 
 Sky-View Factor (102)
 ^^^^^^^^^^^^^^^^^^^^^
@@ -173,7 +167,6 @@ The fraction of visible sky in a given area. **naturf** generalizes the distance
 .. math::
   SVF = cos(arctan(\frac{z_{h}}{0.5W}))
 
-where, `SVF` is the sky-view factor; `H` is the building height in m; `W` is the distance between buildings in m.
 
 Grimmond & Oke Roughness Length (103)
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
[…]
