### 2024-05-03 51c0ce9
fix equations not rendering correctly #164
