### 2024-05-03 3bf513a
fix equations not rendering correctly #164
