### 2024-05-03 97d1a98
fix equations not rendering correctly #164
