### 2024-05-03 IMMM-SFA/naturf#164 @levisweetbreu
A glossary of terms has now been added and terms have been synthesized across equations!
