### 2024-05-04 e54cb27
Updates instructions for logging executions to the Hamilton UI
Added instructions on how to see and log executions
using the Hamilton UI. Links to the correct documentation
to help people get started.
