### 2024-05-04 b5bb7ed
Adds pragma to skip some lines from coverage report
Because we don't want to test them.
