### 2024-05-04 998d0b3
Updates instructions for logging executions to the Hamilton UI (#166)
* Updates instructions for logging executions to the Hamilton UI

Added instructions on how to see and log executions
using the Hamilton UI. Links to the correct documentation
to help people get started.

* Adds unit test for catching SDKs not installed

Adding test to make coverage not complain, and to
also validate that if the environment variables are present
that things wont break if the right SDK is not installed.

* Adds pragma to skip some lines from coverage report

Because we don't want to test them.
