### 2024-05-23 IMMM-SFA/naturf#168 @caimeng2
model.graph() path
For your examples provided in `quickstarter.ipynb`, `model.execute()` works fine, but `model.graph()` doesn't. I encountered

> ExecutableNotFound: failed to execute PosixPath('dot'), make sure the Graphviz executables are on your systems' PATH

It would be great if you could add some instructions on how to set up this path.

reference: https://github.com/openjournals/joss-reviews/issues/6712
