### 2024-05-23 IMMM-SFA/naturf#169 @caimeng2
license inconsistency
The license listed `pyproject.toml` is BSD while the `LICENSE` indicating MIT.

btw, the authors listed `pyproject.toml` seem to be outdated as well.

reference: https://github.com/openjournals/joss-reviews/issues/6712
