### 2024-05-23 IMMM-SFA/naturf#168 @erexer
Hi! A couple questions:
1. Did you `pip install graphviz`?
2. Did you install the requirements as listed in the pyproject.toml?
3. What operating system are you on?
