### 2024-05-23 IMMM-SFA/naturf#168 @caimeng2
1. Yes.
2. You mean installing the dependencies manually? No. I did pip install as the instruction says.  If you think this is the reason for the error, I would suggest handling the requirements automatically (e.g., `requirements.txt`)
3. macOS 13.0.1
