### 2024-05-24 #6712 @caimeng2
Hi,

This is great software. I can see it being useful. And the documentation is sufficient. I have two more suggestions for improvement:

1. The description of the functionality in the paper could be made clearer. Something similar to the "Purpose" section in README would be nice. Currently, the "Design and Functionality" section focuses more on the design aspects rather than explaining the functionality.

2. I noticed a minor issue where the version number in the [user guide](https://immm-sfa.github.io/naturf/user_guide.html) is listed as 1.0.1, while the latest version is 1.0.2. Please ensure that the use guide is up-to-date.

Aside from these, and the three open issues already noted, I am satisfied with the package and would recommend acceptance.
