### 2024-05-24 57d3186
change pyproject.toml to add Stefan to the author list, remove python3.8, and change license to MIT (#170)
