### 2024-05-24 IMMM-SFA/naturf#171 @praneethd7
[JOSS] Minor Suggestions for improvement
I have a couple of suggestions to improve the package and documentation. These are non-binding to acceptance of course.

- The graphs are amazing and give a good idea of the subprocess. Why not make `graphviz` library a dependency in the package? Especially because the tutorial has it at the start
- The quick starter guide could be easily transformed into a Google Colab notebook and you can have a badge for that in the README. This helps users get started more conveniently
- It would be nice to have the GitHub Repo link in the documentation that is easy to access from any page
- Besides computing these parameters it would be nice to have some visualization capabilities (such as Figure 1) built-in using `geopandas` (since you are using it already) or other visualization libraries. Also, would be nice to have tutorials on integrating this work with other use cases mentioned in the manuscript:

> (i) urban climate modelers wanting to understand building23
effects on the urban microclimate at a fine scale, (ii) urban planners creating new developments,24
(iii) sociologists aiming to understand weather-based inequalities and stresses

In the current state of the package, the users are expected to connect the dots in between for their use case😄

- Currently the automated package build tests Linux (ubuntu) and Python 3.10. It would be good to test on more OS and versions of Python to cater to a wider audience
- It would be nice to have some links to open-source/data for obtaining the building polygons and heights data

Cross Reference: openjournals/joss-reviews#6712
