### 2024-05-24 IMMM-SFA/naturf#172 @praneethd7
[JOSS] Manuscript Clarifications and Suggestions
Here are a couple of queries and suggestions on the manuscript:

- It would be nice to have a figure showing the workflow and mentioning the part about selecting the variables of interest within the manuscript
- The `Statement of Need` section can be improved. It is not clear what NATURF does that other existing packages do not. When you started this work, what was lacking in the state of the field that motivated you all to build this? Also, the first paragraph is very repetitive. Please be terse or not repetitive.
- I do not understand how `NATURF` can provide outputs at `level 2` which the `GeoClimate` package cannot. It appears to be an artifact of input data rather than something `NATURF` does. Please clarify here and also in the manuscript.
> NATURF and GeoClimate measure the effect of urban features on microclimate, but they use different source data and output different data products to do so.
- As a reviewer, I also want to know how this package compares to other packages ( GeoClimate and others) in terms of computation time, memory usage, and functionality/ convenience (besides the WUDAPT levels). This helps the check the following box:
> **State of the field:** Do the authors describe how this software compares to other commonly used packages?
- Please clarify this sentence:
> Urban planners will be able to use NATURF in the same way as urban areas continue to grow

### Minor comments:
- The GitHub repository is listed on the left side of the manuscript at the top as per JOSS format. So this seems like a repetition.
> The source code for NATURF can be found on GitHub
- Please add a DOI or permanent link for "An introduction to the WUDAPT project" if possible

Cross Reference: openjournals/joss-reviews#6712
