### 2024-05-24 IMMM-SFA/naturf#168 @erexer
makes sense, I'll add it to the required packages list
