+++ paper/paper.md
@@ -35,19 +35,21 @@ bibliography: paper.bib
 ---
 
 # Summary
-The Neighborhood Adaptive Tissues for Urban Resilience Futures tool (NATURF) is a Python workflow that generates files readable by the Weather Research and Forecasting (WRF) model. NATURF uses *geopandas* [@kelsey_jordahl_2020_3946761] and *hamilton* [@DBLP:conf/vldb/KrawczykI22] to calculate 132 building parameters from shapefiles with building footprint and height information. These parameters can be collected and used in many formats, and the primary output is a binary file configured for input to WRF. This workflow is a flexible adaptation of the National/World Urban Database and Access Portal Tool (NUDAPT/WUDAPT)[@ching2009national; @mills2015introduction] that can be used with any study area at any spatial resolution. The climate modeling community and urban planners can identify the effects of building/neighborhood morphology on the microclimate using the urban parameters and WRF-readable files produced by NATURF. The source code for NATURF can be found on [GitHub](https://github.com/IMMM-SFA/naturf/tree/main), and more information on the urban parameters calculated can be found in the [documentation](https://immm-sfa.github.io/naturf/).
+The Neighborhood Adaptive Tissues for Urban Resilience Futures tool (NATURF) is a Python workflow that generates files readable by the Weather Research and Forecasting (WRF) model. NATURF uses *geopandas* [@kelsey_jordahl_2020_3946761] and *hamilton* [@DBLP:conf/vldb/KrawczykI22] to calculate 132 building parameters from shapefiles with building footprint and height information. These parameters can be collected and used in many formats, and the primary output is a binary file configured for input to WRF. This workflow is a flexible adaptation of the National/World Urban Database and Access Portal Tool (NUDAPT/WUDAPT)[@ching2009national; @mills2015introduction] that can be used with any study area at any spatial resolution. The climate modeling community and urban planners can identify the effects of building/neighborhood morphology on the microclimate using the urban parameters and WRF-readable files produced by NATURF. More information on the urban parameters calculated can be found in the [documentation](https://immm-sfa.github.io/naturf/).
 
 # Statement of Need
-NATURF serves many audiences: (i) urban climate modelers wanting to understand building effects on the urban microclimate at a fine scale, (ii) urban planners creating new developments, (iii) sociologists aiming to understand weather-based inequalities and stresses. A preliminary version of NATURF was used to calculate urban parameters and demonstrate that simulated new developments in the Chicago Loop neighborhood in Chicago Illinois, USA affect temperature and energy use both in the new developments and the preexisting neighborhoods [@allen2020impacts]. Their findings show that building effects on the microclimate can be modeled at 90m resolution, and they quantify how different configurations of urban developments affect not only the developments themselves but also neighborhoods that already exist. Urban planners will be able to use NATURF in the same way as urban areas continue to grow. Likewise, NATURF will give climate modelers the tools to understand how urbanization will contribute to microclimate and broader global climate change, and sociologists could see how urban developments affect weather-related stresses.
[…]
