+++ paper/paper.bib
@@ -26,7 +26,8 @@
   author={Mills, Gerald and Ching, J and See, L and Bechtel, B and Foley, M},
   booktitle={Proceedings of the 9th International Conference on Urban Climate, Toulouse, France},
   pages={20--24},
-  year={2015}
+  year={2015},
+  url={https://www.wudapt.org/wp-content/uploads/2015/05/Mills_etal_ICUC9.pdf}
 }
 @article{ching2018wudapt,
   title={WUDAPT: An urban weather, climate, and environmental modeling infrastructure for the anthropocene},
@@ -104,4 +105,28 @@
   version      = {v0.8.1},
   doi          = {10.5281/zenodo.3946761},
   url          = {https://doi.org/10.5281/zenodo.3946761}
+}
+
+@article{best2006progress,
+  title={Progress towards better weather forecasts for city dwellers: from short range to climate change},
+  author={Best, MJ},
+  journal={Theoretical and Applied Climatology},
+  volume={84},
+  pages={47--55},
+  year={2006},
+  publisher={Springer},
+  doi={10.1007/s00704-005-0143-2},
+  url={https://doi.org/10.1007/s00704-005-0143-2}
+}
+
+@article{wong2019evaluation,
+  title={Evaluation of uWRF performance and modeling guidance based on WUDAPT and NUDAPT UCP datasets for Hong Kong},
+  author={Wong, Michael Mau Fung and Fung, Jimmy Chi Hung and Ching, Jason and Yeung, Peter Pak Shing and Tse, Jason Wai Po and Ren, Chao and Wang, Ran and Cai, Meng},
+  journal={Urban Climate},
+  volume={28},
+  pages={100460},
+  year={2019},
+  publisher={Elsevier},
+  doi={10.1016/j.uclim.2019.100460},
+  url={https://doi.org/10.1016/j.uclim.2019.100460}
 }
