+++ paper/paper.bib
@@ -130,3 +130,12 @@
   doi={10.1016/j.uclim.2019.100460},
   url={https://doi.org/10.1016/j.uclim.2019.100460}
 }
+
+@software{modelamerica,
+  title={Model America: Data and Models for every U.S. Building},
+  author={New, Joshua and Bass, Brett and Berres, Andy and Clinton, Nicholas and Adams, Mark and Li, Fengqi and Stubbings, Avery and Chowdhury, Shovan},
+  year={2021},
+  publisher={ESS-DIVE repository},
+  doi={10.15485/2283980},
+  url={https://data.ess-dive.lbl.gov/datasets/doi:10.15485/2283980}
+}
