+++ paper/paper.md
@@ -38,12 +38,12 @@ bibliography: paper.bib
 The Neighborhood Adaptive Tissues for Urban Resilience Futures tool (NATURF) is a Python workflow that generates files readable by the Weather Research and Forecasting (WRF) model. NATURF uses *geopandas* [@kelsey_jordahl_2020_3946761] and *hamilton* [@DBLP:conf/vldb/KrawczykI22] to calculate 132 building parameters from shapefiles with building footprint and height information. These parameters can be collected and used in many formats, and the primary output is a binary file configured for input to WRF. This workflow is a flexible adaptation of the National/World Urban Database and Access Portal Tool (NUDAPT/WUDAPT)[@ching2009national; @mills2015introduction] that can be used with any study area at any spatial resolution. The climate modeling community and urban planners can identify the effects of building/neighborhood morphology on the microclimate using the urban parameters and WRF-readable files produced by NATURF. More information on the urban parameters calculated can be found in the [documentation](https://immm-sfa.github.io/naturf/).
 
 # Statement of Need
-Fine-resolution representations of urban surfaces are necessary to support weather modeling at informative scales for understanding urban microclimate effects on public health and energy use [@best2006progress]. NUDAPT/WUDAPT were both created to address this need, but data is only currently available for select regions [@ching2009national] and using inputs such as satellite data [@ching2018wudapt] that could introduce uncertainties in the parameters produced [wong2019evaluation]. NATURF uses building-level data to represent urban surfaces at sub-kilometer resolutions, allowing the user to conduct studies at a high resolution at any location where building footprint and height data exist. A preliminary version of NATURF was used to calculate urban parameters and demonstrate that simulated new developments in the Chicago Loop neighborhood in Chicago Illinois, USA affect temperature and energy use both in the new developments and the preexisting neighborhoods [@allen2020impacts]. Their findings show that building effects on the microclimate can be modeled at 90m resolution, and they quantify how different configurations of urban developments affect not only the developments themselves but also neighborhoods that already exist. In this way, urban planners will be able to use NATURF to project how urban developments will affect meterology within and outside of the new growth.
+Fine-resolution, three-dimensional representations of urban surfaces are necessary to support weather modeling at informative scales for understanding urban microclimate effects on public health and energy use [@best2006progress]. NUDAPT/WUDAPT were both created to address this need, but data is only currently available for select regions [@ching2009national] and using inputs such as satellite data [@ching2018wudapt] that could introduce uncertainties in the parameters produced [wong2019evaluation]. NATURF uses building-level data to represent urban surfaces at sub-kilometer resolutions, allowing the user to conduct studies at a high resolution at any location where building footprint and height data exist. A preliminary version of NATURF was used to calculate urban parameters and demonstrate that simulated new developments in the Chicago Loop neighborhood in Chicago Illinois, USA affect temperature and energy use both in the new developments and the preexisting neighborhoods [@allen2020impacts]. Their findings show that building effects on the microclimate can be modeled at 90m resolution, and they quantify how different configurations of urban developments affect not only the developments themselves but also neighborhoods that already exist. In this way, urban planners will be able to use NATURF to project how urban developments will affect meteorology within and outside of the new growth.
 
[…]
