+++ paper/paper.md
@@ -43,7 +43,7 @@ Fine-resolution, three-dimensional representations of urban surfaces are necessa
 Similar to NATURF, the open-source toolbox GeoClimate [@bocher2021geoclimate] aims to quantify the effect of urban features for climate models. Both tools provide outputs for WRF to be used with the NUDAPT/WUDAPT framework, but at different resolutions. GeoClimate is intended for the production of local climate zones, data classified as level 0 by WUDAPT used to approximate urban parameters within the zone [@ching2018wudapt]. NATURF produces level 2 data, precise urban parameter data at finer resolutions. The input data used by each tool influences the resolution of outputs; GeoClimate primarily uses OpenStreetMap data where building heights must be estimated while NATURF uses shapefiles with building height and footprint data included. For both tools, study areas at the neighborhood-scale have comparable memory usage and computational time. For larger areas, particularly at the city-scale, NATURF is more suited for use on high-performance computers (HPC) due to the number and complexity of interactions being calculated. Despite this, installation and execution of NATURF requires only a few minutes and edits to code, and the use of *hamilton* allows for the tracking of code execution progress. Additionally, NATURF can internally produce outputs ready for WRF whereas GeoClimate requires the installation and use of additional packages. While NATURF and GeoClimate both produce urban parameter data for urban microclimate modeling, NATURF does so at a higher resolution with flexibility to analyze complex interactions between buildings at multiple scales.
 
 # Design and Functionality
-NATURF uses *hamilton* for organization and visualization of its workflow. Function names become inputs to other functions, allowing for a delineation of dependencies for every function. This code format allows for easier troubleshooting and visualization of the workflow (See \autoref{fig{dag}}). Additionally, *hamilton* allows for the calculation of a specified subset of variables if the entire workflow is not desired. The calculation of parameters itself is done through *geopandas*. The input shapefile is loaded in as a GeoDataFrame that can handle calculation of parameters for each building in one process as opposed to using long "for" or "while" loops. Likewise, the geospatial features of *geopandas* are conducive to calculating the urban parameters for NATURF which are based on the geometry of each building and its neighbors. \autoref{fig:bldg_ht} shows one of these parameters, mean building height at 100 meter resolution over Los Angeles County, California, USA using open-source building data from ModelAmerica [@modelamerica].
+NATURF uses *hamilton* for organization and visualization of its workflow. Function names become inputs to other functions, allowing for a delineation of dependencies for every function. This code format allows for easier troubleshooting and visualization of the workflow (See \autoref{fig{dag}}). Additionally, *hamilton* allows for the calculation of a specified subset of variables if the entire workflow is not desired. The calculation of parameters itself is done through *geopandas*. The input shapefile is loaded in as a GeoDataFrame that can handle calculation of parameters for each building in one process as opposed to using long "for" or "while" loops. Likewise, the geospatial features of *geopandas* are conducive to calculating the urban parameters for NATURF which are based on the geometry of each building and its neighbors. \autoref{fig:bldg_ht} shows one of these parameters, mean building height at 100 meter resolution over Los Angeles County, California, USA using open-source building data from Model America [@modelamerica].
 
 ![NATURF workflow visualization. Desired variables can be specified as inputs and all dependent functions will be executed. \label{fig:dag}](figure_1.png)
 
