### 2024-05-28 #6712 @levisweetbreu
@caimeng2 Thank you for your thoughtful review and comments! To your first suggestion above, I have added a new paragraph to the "Design and Functionality" section to better describe the functionality of our software.
