### 2024-05-28 IMMM-SFA/naturf#172 @levisweetbreu
@praneethd7 Thank you for your suggestions! Edits to the manuscript have been made in the last series of pull requests and commits.

To address your specific suggestions:

> It would be nice to have a figure showing the workflow and mentioning the part about selecting the variables of interest within the manuscript

A new figure showing the workflow (now Figure 1) has been added as well as a line discussing the ability for users to select variables of interest from the workflow.

> The Statement of Need section can be improved. It is not clear what NATURF does that other existing packages do not. When you started this work, what was lacking in the state of the field that motivated you all to build this? Also, the first paragraph is very repetitive. Please be terse or not repetitive.

The `Statement of Need` section has been reworked to better clarify the scientific gap addressed by `NATURF`. The first paragraph of the section now addresses the motivations to create `NATURF` and the needs filled by the software. The remainder of the paragraph was edited to reduce repetition.

> I do not understand how NATURF can provide outputs at level 2 which the GeoClimate package cannot. It appears to be an artifact of input data rather than something NATURF does. Please clarify here and also in the manuscript.

The second paragraph of the `Statement of Need` section was reorganized and edited to better compare `GeoClimate` and `NATURF` and clarify the differences in workflow and outputs. `NATURF` produces finer resolution output data, classified as `level 2` both due to the building data input and the specific parameters calculated that focus on interactions between buildings. `GeoClimate` produces coarser local climate zones that approximate urban parameters within a given grid cell, resulting in `level 0` data.

> As a reviewer, I also want to know how this package compares to other packages ( GeoClimate and others) in terms of computation time, memory usage, and functionality/ convenience (besides the WUDAPT levels). This helps the check the following box:

A few lines on the requested comparisons have been added. `NATURF` and `GeoClimate` have comparable computation time and memory usage at neighborhood-scales, but `NATURF` requires more time and memory at city-scale due to the complex interactions being calculated. In terms of functionality/convenience, `NATURF` more easily creates outputs usable in WRF and offers flexibility in parameters calculated as well as visualization and tracking of code execution.

> Please clarify this sentence: Urban planners will be able to use NATURF in the same way as urban areas continue to grow

This sentence has been reworded: "In this way, urban planners will be able to use NATURF to project how urban developments will affect meteorology within and outside of the new growth."

> The GitHub repository is listed on the left side of the manuscript at the top as per JOSS format. So this seems like a repetition.

The indicated link has been removed.

> Please add a DOI or permanent link for "An introduction to the WUDAPT project" if possible

A link has been added to the citation.
