### 2024-05-28 f146209
add reference and fix figure reference
