+++ paper/paper.bib
@@ -139,3 +139,12 @@
   doi={10.15485/2283980},
   url={https://data.ess-dive.lbl.gov/datasets/doi:10.15485/2283980}
 }
+
+@software{urbparams,
+  title={Urban Parameters LA County 100m},
+  author={Sweet-Breu, Levi and Allen-Dumas, Melissa},
+  year={2024},
+  publisher={MSD-Live Data Repository},
+  doi={10.57931/2349436},
+  url={https://doi.org/10.57931/2349436}
+}
