### 2024-05-29 #6712 @praneethd7
@levisweetbreu @cheginit Hello. I have completed the review from my end and I am pleased to accept this submission for publication😄. The authors have patiently addressed all of my comments🤗. This is a great package that will be helpful to researchers trying to compute these urban parameters and model built environments. The package does exactly what it promises and installation (and usage) was straightforward. The documentation, code formatting, installation instructions, and automated tests are all in place and well-written, thereby encouraging community contributions. Happy to have reviewed this work!

I do have a set of minor suggestions but I leave it to the author's discretion to address them
- [naturf/issues#171](https://github.com/IMMM-SFA/naturf/issues/171#issue-2315992242)
- The new Figure 1 has a small font size that might be hard for some readers. I suggest increasing it and/or having the figure as PDF so that the vector image does not lose resolution upon scaling it ☺️
