### 2024-05-29 dc17bb1
Update driver.py to use inputs (#174)
Inputs can be passed in as configuration, but really they should be passed in as inputs -- that way visualization doesn't interpret them as configuration.
