### 2024-05-29 17bd245
update dependencies, wording on quickstarter
