+++ docs/source/quickstarter.rst
@@ -8,7 +8,10 @@ Input data
 
 The only input data required for `naturf` is a shapefile with building footprints and height data. There should be a field with a unique ID for each building the shapefile, and it should be in a projected coordinate system such as Alber Equal Area Conic. For input to the Weather Research and Forecasting model (WRF), the computed parameters for each building will be projected into WGS 84.
 
-Either check out our interactive `quickstarter <quickstarter.rst>`_, or continue below to run `naturf` using a python file. Users need to have `graphviz <https://graphviz.org/download/>`_ installed to visualize the Directed Acyclic Graph (DAG).
+Either run the ipynb `quickstarter <https://github.com/IMMM-SFA/naturf/blob/main/notebooks/quickstarter.ipynb>`_, or continue below to run `naturf` using a python file.
+
+.. note::
+    Users need to `download graphviz <https://graphviz.org/download/>`_ to visualize the Directed Acyclic Graph (DAG).
 
 1. Install `naturf`
 -------------------
