+++ pyproject.toml
@@ -28,19 +28,19 @@ classifiers = [
 ]
 
 dependencies = [
-  "fiona>=1.8.19",
-  "geocube>=0.3.1",
-  "geopandas>=0.10.2",
-  "joblib>=1.0.1",
-  "numpy>=1.22.4",
-  "pandas>=1.4.2",
+  "fiona>=1.9.6",
+  "geocube>=0.5.2",
+  "geopandas>=0.14.4",
+  "joblib>=1.4.2",
+  "numpy>=1.26.4",
+  "pandas>=2.2.2",
   "pyproj>=3.0.1",
-  "rasterio>=1.2.10",
-  "rtree>=1.0.0",
-  "sf-hamilton[visualization]>=1.45",
-  "shapely>=1.8.2, <2",
-  "tqdm>=4.66.1",
-  "xarray>=2022.3.0",
+  "rasterio>=1.3.10",
+  "rtree>=1.2.0",
+  "sf-hamilton[visualization]>=1.64",
+  "shapely>=2.0.4",
+  "tqdm>=4.66.4",
+  "xarray>=2024.5.0",
 ]
 
 [project.optional-dependencies]
