+++ pyproject.toml
@@ -7,10 +7,10 @@ name = "naturf"
 version = "1.0.2"
 description = 'The Neighborhood Adaptive Tissues for Urban Resilience Futures tool (NATURF) is a Python workflow that generates data readable by the Weather Research and Forecasting (WRF) model. The NATURF Python modules use shapefiles containing building footprint and height data as input to calculate 132 building parameters at any resolution and converts the parameters into a binary file format.'
 readme = "README.md"
-requires-python = ">=3.8"
+requires-python = ">=3.9"
 keywords = []
 authors = [
-  { name = "Em Rexer", email = "emily.rexer@pnnl.gov" },
+  { name="Em Rexer", email="emily.rexer@pnnl.gov" },
   { name="Levi Sweet-Breu", email="levi_sweetbreu1@baylor.edu" },
   { name="Melissa Allen-Dumas", email="allenmr@ornl.gov" },
   { name="Chris Vernon", email="chris.vernon@pnnl.gov" },
@@ -28,19 +28,26 @@ classifiers = [
 ]
 
 dependencies = [
-  "fiona>=1.9.6",
-  "geocube>=0.5.2",
-  "geopandas>=0.14.4",
+  "fiona==1.8.19;python_version<'3.10'",
+  "fiona>=1.9.6;python_version>='3.10'",
+  "geocube==0.3.3;python_version<'3.10'",
+  "geocube>=0.5.2;python_version>='3.10'",
+  "geopandas==0.10.2;python_version<'3.10'",
+  "geopandas>=0.14.4;python_version>='3.10'",
+  "graphviz==0.20.3",
   "joblib>=1.4.2",
-  "numpy>=1.26.4",
+  "numpy==1.22.4;python_version<'3.10'",
+  "numpy>=1.26.4;python_version>='3.10'",
   "pandas>=2.2.2",
   "pyproj>=3.0.1",
   "rasterio>=1.3.10",
   "rtree>=1.2.0",
-  "sf-hamilton[visualization]>=1.64",
+  "sf-hamilton[visualization]==1.64;python_version<'3.10'",
+  "sf-hamilton[visualization]>=1.64;python_version>='3.10'",
   "shapely>=2.0.4",
   "tqdm>=4.66.4",
-  "xarray>=2024.5.0",
+  "xarray==2022.3.0;python_version<'3.10'",
+  "xarray>=2024.5.0;python_version>='3.10'",
 ]
 
 [project.optional-dependencies]
