### 2024-05-30 2d2b47d
Bugfix graphviz (#175)
* update dependencies, wording on quickstarter

* update dependencies
