### 2024-06-02 #6712 @cheginit
@levisweetbreu There are two references without DOI, and it appears that they don't have one. Please confirm the missing DOIs.

Once confirmed, please do the following, so I can then move forward with recommending acceptance of the submission:

- [ ] Archive the reviewed software in Zenodo or a similar service (e.g., figshare, an institutional repository). Make sure that the license listed for the archive is the same as the software license!
