+++ paper/paper.bib
@@ -59,7 +59,7 @@
   title     = {Hamilton: a modular open source declarative paradigm for high level
                modeling of dataflows},
   booktitle = {1st International Workshop on Composable Data Management Systems,
-               CDMS@VLDB 2022, Sydney, Australia, September 9, 2022},
+               CDMS@VLDB 2022, {Sydney, Australia}, {September} 9, 2022},
   year      = {2022},
   url       = {https://cdmsworkshop.github.io/2022/Proceedings/ShortPapers/Paper6\_StefanKrawczyk.pdf},
   timestamp = {Wed, 19 Oct 2022 16:20:48 +0200},
@@ -120,7 +120,7 @@
 }
 
 @article{wong2019evaluation,
-  title={Evaluation of uWRF performance and modeling guidance based on WUDAPT and NUDAPT UCP datasets for Hong Kong},
+  title={Evaluation of uWRF performance and modeling guidance based on WUDAPT and NUDAPT UCP datasets for {Hong Kong}},
   author={Wong, Michael Mau Fung and Fung, Jimmy Chi Hung and Ching, Jason and Yeung, Peter Pak Shing and Tse, Jason Wai Po and Ren, Chao and Wang, Ran and Cai, Meng},
   journal={Urban Climate},
   volume={28},
@@ -132,7 +132,7 @@
 }
 
 @software{modelamerica,
-  title={Model America: Data and Models for every U.S. Building},
+  title={Model America: Data and Models for every {U.S.} Building},
   author={New, Joshua and Bass, Brett and Berres, Andy and Clinton, Nicholas and Adams, Mark and Li, Fengqi and Stubbings, Avery and Chowdhury, Shovan},
   year={2021},
   publisher={ESS-DIVE repository},
@@ -141,7 +141,7 @@
 }
 
 @software{urbparams,
-  title={Urban Parameters LA County 100m},
+  title={Urban Parameters LA {County} 100m},
   author={Sweet-Breu, Levi and Allen-Dumas, Melissa},
   year={2024},
   publisher={MSD-Live Data Repository},
