### 2024-06-05 #6712 @levisweetbreu
@cheginit I confirm that there are not DOIs for those two references.

Our version is:  v1.0.3
Our archive DOI is:  10.5281/zenodo.11454159
