### 2024-06-06 #6712 @kthyng
Paper comments:
* Need a space in the first paragraph "(NUDAPT/WUDAPT)(Ching et al., 2009; Mills et al., 2015)"
* Statement of Need first paragraph "Chicago Illinois, USA" needs a comma after Chicago
* Please check the capitalization in your references. You can preserve capitalization by placing {} around characters/words in your .bib file. "anthropocene" is not capitalized nor "america".
