+++ paper/paper.bib
@@ -30,7 +30,7 @@
   url={https://www.wudapt.org/wp-content/uploads/2015/05/Mills_etal_ICUC9.pdf}
 }
 @article{ching2018wudapt,
-  title={WUDAPT: An urban weather, climate, and environmental modeling infrastructure for the anthropocene},
+  title={WUDAPT: An urban weather, climate, and environmental modeling infrastructure for the {Anthropocene}},
   author={Ching, Jason and Mills, G and Bechtel, Benjamin and See, L and Feddema, J and Wang, X and Ren, C and Brousse, O and Martilli, A and Neophytou, M and others},
   journal={Bulletin of the American Meteorological Society},
   volume={99},
@@ -132,7 +132,7 @@
 }
 
 @software{modelamerica,
-  title={Model America: Data and Models for every {U.S.} Building},
+  title={Model {America}: Data and Models for every {U.S.} Building},
   author={New, Joshua and Bass, Brett and Berres, Andy and Clinton, Nicholas and Adams, Mark and Li, Fengqi and Stubbings, Avery and Chowdhury, Shovan},
   year={2021},
   publisher={ESS-DIVE repository},
